fix(bulk): add submit for bulk, add more test, add mock-api for upload-session
