+++ src/features/bulk/steps/Step2.tsx
@@ -1,35 +1,44 @@
-import React from 'react';
+import React, { useCallback, useState } from 'react';
 import { Navigate } from 'react-router-dom';
 
-import { MetadataGenericType } from '../../../types';
+import { CustomJsonSchema, MetadataGenericType } from '../../../types';
 import BulkMetadataForm from '../../../components/BulkMetadataForm/BulkMetadataForm';
 import { STEP1 } from '../bulk/constants';
 import { getFilesFromStore } from '../bulk/selectors';
 import { Props } from '../bulk/types';
 import BulkWizard from '../wizard/BulkWizard';
-import { useAppSelector } from '../../hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { setFieldData } from '../bulk/slice';
 
 export interface Step2Props<T> extends Props<T> {
-	localData: any;
-	onChange: (data: MetadataGenericType, valid: boolean) => void;
 }
 
 export default function Step2<T>(props: Step2Props<T>) {
-	const { metadataForm, localData, onChange } = props;
+	const { metadataForm, metadataFields } = props;
+
+	const dispatch = useAppDispatch();
+	const [localData, setLocalData] = useState<CustomJsonSchema>(metadataForm.data)
 	const filesFromStore = useAppSelector(getFilesFromStore);
+	const [isValidForm, setIsValidForm] = useState<boolean>(false);
+
+	const handleOnChange = useCallback((data: MetadataGenericType, valid: boolean) => {
+		dispatch(setFieldData(data))
+		setLocalData(data);
+		setIsValidForm(valid);
+	}, [metadataFields, metadataForm])
 
 	// Redirect to step1 when state is not correct
 	if (filesFromStore?.length === 0) {
 		return <Navigate to={STEP1} />;
 	} else {
 		return (
-			<BulkWizard {...props}>
+			<BulkWizard {...props} isValidForm={isValidForm}>
 				<BulkMetadataForm
 					schema={metadataForm.schema}
 					uischema={metadataForm.uischema}
 					renderers={metadataForm.renderers}
 					data={localData}
-					onChange={onChange}
+					onChange={handleOnChange}
 				/>
 			</BulkWizard>
 		);
